Reorganize source folders in swf combined editor
