+++ packages/serverless-logic-web-tools/src/editor/channel/SwfChannelComponent.tsx
@@ -35,7 +35,7 @@ import { SwfLanguageServiceChannelApiImpl } from "../api/SwfLanguageServiceChann
 import { SwfServiceCatalogChannelApiImpl } from "../api/SwfServiceCatalogChannelApiImpl";
 import { WebToolsSwfLanguageService } from "../api/WebToolsSwfLanguageService";
 import { EditorChannelComponentProps, EditorChannelComponentRef } from "../WebToolsEmbeddedEditor";
-import { SwfCombinedEditorChannelApiImpl } from "@kie-tools/serverless-workflow-combined-editor/dist/impl";
+import { SwfCombinedEditorChannelApiImpl } from "@kie-tools/serverless-workflow-combined-editor/dist/channel";
 import { MessageBusClientApi } from "@kie-tools-core/envelope-bus/dist/api";
 import { Notification } from "@kie-tools-core/notifications/dist/api";
 import { Position } from "monaco-editor";
