+++ packages/serverless-workflow-combined-editor/src/channel/SwfStaticEnvelopeContentProviderChannelApiImpl.ts
