+++ packages/serverless-workflow-combined-editor/src/envelope/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2023 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,5 +14,4 @@
  * limitations under the License.
  */
 
-export * from "./editor";
-export * from "./impl";
+export * from "./ServerlessWorkflowCombinedEditorEnvelopeApiImpl";
